feat: add config system and build binary
- Added JSON configuration system with embedded config file
- Configurable window size, position alignment, colors
- Implemented uniform color system for text rendering
- Built production binary for immediate use
- Maintained all existing functionality with config flexibility
- Changed the license from AGPL to MIT
- The project name has been changed to TimeChecker

+++ internal/ui/btns.go
@@ -32,10 +32,10 @@ func (b *btn) setText(t string) {
 	b.Text = t
 }
 
-func (b *btn) Hover(w *glfw.Window, winW, winH float32) (bool) {
+func (b *btn) Hover(w *glfw.Window, winW, winH int) (bool) {
 	mX, mY := w.GetCursorPos()
-	glX := float32(mX)/winW*2-1
-	glY := 1-float32(mY)/winH*2
+	glX := float32(mX)/float32(winW)*2-1
+	glY := 1-float32(mY)/float32(winH)*2
 
 	if glX >= b.pos[0] && glX <= b.pos[2] && glY <= b.pos[1] && glY >= b.pos[3] {
 			return true
